continued script

+++ common/accuracy_evaluator/src/python/fileHandling.py
@@ -1,13 +1,16 @@
-from shutil import copyfile
-from in_gen import in_gen
-from samples_generator import gen_samples
+import shutil
+import settings as s
+import isa
+import os
+import subprocess
 
 def removeTree(root):
-	shutil.rmtree(root)
+	if os.path.exists(root):
+		shutil.rmtree(root)
 
 def mkdir(path):
 	if not os.path.exists(path):
-    	os.makedirs(path)
+		os.makedirs(path)
 
 def cleanServer(session):
 	session.clean(s.remote_srcPath)
@@ -26,15 +29,15 @@ def buildTree():
 	mkdir(s.tmpPath)
 	mkdir(s.reportPath)
 	mkdir(s.resultPath)
+	mkdir(s.scriptPath)
 
 def uploadScripts(session):
-	session.copy_to(s.mult_tcl_name, s.remote_simPath)
-	session.copy_to(s.filter_tcl_name, s.remote_simPath)
-	session.copy_to(s.sMult_tcl_name, s.remote_simPath)
-	session.copy_to(s.sFilter_tcl_name, s.remote_simPath)
+	session.copy_to(s.mult_sim_tcl_name, s.remote_simPath)
+	session.copy_to(s.filter_sim_tcl_name, s.remote_simPath)
+	session.copy_to(s.sMult_sim_tcl_name, s.remote_simPath)
+	session.copy_to(s.sFilter_sim_tcl_name, s.remote_simPath)
 	session.copy_to(s.mult_synth_tcl_name, s.remote_synPath)
 	session.copy_to(s.filter_synth_tcl_name, s.remote_synPath)
-	session.copy_to(s.report_tcl_name, s.remote_synPath)
 
 def uploadSamples(session):
 	samplesPath = '{}/{}'.format(s.inputPath, s.multSamples_name)
@@ -44,58 +47,58 @@ def uploadSamples(session):
 
 def uploadSrcTb(session):
 	# multiplier related files
-	for vhdFile in os.listdir(s.local_srcMult)
+	for vhdFile in os.listdir(s.local_srcMult):
 		if (vhdFile.endswith('.vhd')):
 			if (vhdFile != '{}.vhd'.format(s.multBaseName)): 
 				filePath = '{}/{}'.format(s.local_srcMult, vhdFile)
-				session.copy_to(filePath, remote_v3Path)
+				session.copy_to(filePath, s.remote_v3Path)
 
 	# project specific files
-	for vhdFile in os.listdir(s.local_project)
+	for vhdFile in os.listdir(s.local_project):
 		if (vhdFile.endswith('.vhd')):
 				filePath = '{}/{}'.format(s.local_project, vhdFile)
-				session.copy_to(filePath, remote_srcPath)
+				session.copy_to(filePath, s.remote_srcPath)
 
 	# generic src files
-	for vhdFile in os.listdir(s.local_src)
+	for vhdFile in os.listdir(s.local_src):
 		if (vhdFile.endswith('.vhd')):
 			if (vhdFile != '{}.vhd'.format(s.multBaseName)): 
-				filePath = '{}/{}'.format(s.local_srcMult, vhdFile)
-				session.copy_to(filePath, remote_srcPath)
+				filePath = '{}/{}'.format(s.local_src, vhdFile)
+				session.copy_to(filePath, s.remote_srcPath)
 
 	# filter tb files
-	for vhdFile in os.listdir(s.local_tb)
-		if (vhdFile.endswith('.vhd')):
-				filePath = '{}/{}'.format(s.local_tbMult, vhdFile)
-				session.copy_to(filePath, remote_tbPath)
+	for vhd_v_File in os.listdir(s.local_tb):
+		if (vhd_v_File.endswith('.vhd') or vhd_v_File.endswith('.v')):
+				filePath = '{}/{}'.format(s.local_tb, vhd_v_File)
+				session.copy_to(filePath, s.remote_tbPath)
 
 	# mult tb files
-	for vhdFile in os.listdir(s.local_tbMult)
+	for vhdFile in os.listdir(s.local_tbMult):
 		if (vhdFile.endswith('.vhd')):
 				filePath = '{}/{}'.format(s.local_tbMult, vhdFile)
-				session.copy_to(filePath, remote_tbPath)
+				session.copy_to(filePath, s.remote_tbPath)
 
 def uploadMultiplier(session):
-	session.copy_to(s.multBaseName, remote_srcPath)
+	session.copy_to(s.multFileName, s.remote_v3Path)
 
 def storeResultsReports(session, compressionLevel, startingDirection):
 	for file in s.resultFilesList:
 		oldResultFilePath = '{}/{}'.format(s.remote_commonPath, file)
 		resultFilePath = '{}/{}{}_{}'.format(s.remote_commonPath, compressionLevel, startingDirection, file)
-		isa.run_commands('mv {oldFile} {newFile}'.format())
+		session.run_commands('mv {oldFile} {newFile}'.format(oldFile=oldResultFilePath, newFile=resultFilePath))
 		session.copy_from(resultFilePath, s.resultPath)
 	for file in s.reportFilesList:
 		oldResultFilePath = '{}/{}'.format(s.remote_commonPath, file)
 		resultFilePath = '{}/{}{}_{}'.format(s.remote_commonPath, compressionLevel, startingDirection, file)
-		isa.run_commands('mv {oldFile} {newFile}'.format())
+		session.run_commands('mv {oldFile} {newFile}'.format(oldFile=oldResultFilePath, newFile=resultFilePath))
 		session.copy_from(resultFilePath, s.reportPath)
 
 def cleanCommon(session):
 	session.clean(s.remote_commonPath)
-	with isa.cd(s.inputPath)
-		for file in listdir('./')
-			if (file.endswith('.txt'))
+	with isa.cd(s.inputPath):
+		for file in os.listdir('./'):
+			if (file.endswith('.txt')):
 				session.copy_to(file, s.remote_commonPath)
 
 def cleanSyn(session):
-	session.clean(remote_synPath)
+	session.clean(s.remote_synPath)
